Added QR scan button and fixed warnings
